requirements.txt
selenium
undetected_chromedriver
numpy
pillow
opencv-python